Global, Hypr, Systems, Flake update 2024-10-21
Refactored systems to be dependant on host. So that jibril and onizuka
can have separate profiles. Doing this split up the systems/ folder, and
also forced a flake update to test. This was done midway through the
hypr rework, and would have been difficult to decouple, so I just did a
big commit instead <3

Refactored the entire hypr system to use settings and nix rather than
its native config. This is working and tested on Jibril, but is missing
the scripts used for switching workspaces on Onizuka.