System, dynamic users for host, removes user.name
adds a new specialArg to hosts: usernameList, which is a list of users
determined through user/user.nix in the host/hostname dir.

This touches syncthing, tuigreet, jibril, shells, and system

removes the user.name field, and converts user.timezone to
system.timezone. This is to move user content off of the host.

This old system is replaced with dynamic user creation based on the
usernameList variable

Because this removes the users.users.defaultUser shells now use the
(more appropriate) users.defaultUserShell