Hypr, fix rebuild not working due to missing icons
The old .gitignore contained the hypr icons path. Newer versions of nix
did not want to rebuild because hypr/icons didn't exist in git. The fix
was simple to move the .gitignore from the home directory to inside of
the icons folder and include it. That way icons can be added and the
folder exists.

Oh--this also adds musescore because that was the impetus to fix this...
It has nothing critical to do with the commit, I'd just rather have them
in one rn