Forgejo, init with admin account and server config
Automatically creates an admin account with sops data. Not my favorite
process because user needs to know to change the sops password to be
owned by forgejo, otherwise good though.

Has a bunch of useful config settings for setting up a server easily.
Commit with that to follow.

Point of worry: SSH_PORT setting isn't set, and will probably need to be
set later to allow people to commit and similar to the repos.