Fish, fix pull/push prompt and change characters
The prompt has been broken for a while now on pull/push, this is because
it was using unicode characters that were larger in the terminal,
causing wraps that would reprint lines. I changed the characters for
push/pull because of it.

I also removed the stash character, and changed the dirty and staged
characters to match the new characters